A few amendments to #341 to make it a bit faster under `Hash_collisions`

### Benchmarks

#### Main

| Method          | Mean     | Error    | StdDev   | Code Size |
|---------------- |---------:|---------:|---------:|----------:|
| Hash_collisions | 13.94 us | 0.105 us | 0.093 us |   3,462 B |

#### SlottedArray

| Method          | Mean     | Error    | StdDev   | Code Size |
|---------------- |---------:|---------:|---------:|----------:|
| Hash_collisions | 14.54 us | 0.147 us | 0.123 us |   3,630 B |

#### This PR

| Method          | Mean     | Error    | StdDev   | Code Size |
|---------------- |---------:|---------:|---------:|----------:|
| Hash_collisions | 13.99 us | 0.111 us | 0.098 us |   3,389 B |